creator-mvp-starter/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3
cryptography==43.0.1
psycopg[binary]==3.2.1
pydantic==2.9.2
